Publicize web view progress constants.
Currently, a developer that wants to execute code when the DOM in a web
view has become interactive must write the following:

```objc
-(void)webViewProgress:(NJKWebViewProgress *)webViewProgress
        updateProgress:(float)progress
{
    if (progress == 0.5) { /* ... */ }
}
```

In the above sample, `0.5` is a magic number.

Eliminate the need to use magic numbers by publicizing the constants
used internally: `0.1`, `0.5`, and `0.9`. This allows users to write the
following:

```objc
if (progress == NJKInteractiveProgressValue) { /* ... */ }
```

Add an explanation of the new constants to the README.